typed-arrays/bigint: fixes per recommendation. Closes gh-1467

+++ test/built-ins/TypedArrays/ctors-bigint/typedarray-arg/custom-proto-access-throws.js
@@ -38,7 +38,7 @@ Object.defineProperty(newTarget, "prototype", {
   }
 });
 
-var sample = new Int8Array();
+var sample = new BigInt64Array();
 
 testWithBigIntTypedArrayConstructors(function(TA) {
   assert.throws(Test262Error, function() {
